requirements.txt
absl-py==0.5.0
adal==1.1.0
alembic==0.8.10
apache-airflow==1.10.0
asn1crypto==0.24.0
astor==0.7.1
Babel==2.6.0
backcall==0.1.0
bleach==2.1.2
boto3==1.9.5
botocore==1.12.5
cachetools==2.1.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
configparser==3.5.0
croniter==0.3.25
cryptography==2.3.1
cycler==0.10.0
cymem==1.31.2
cytoolz==0.9.0.1
decorator==4.3.0
defusedxml==0.5.0
dill==0.2.8.2
docker==3.5.0
docker-pycreds==0.3.0
docutils==0.14
entrypoints==0.2.3
Flask==0.12.4
Flask-Admin==1.4.1
Flask-AppBuilder==1.11.1
Flask-Babel==0.11.1
Flask-Cache==0.13.1
Flask-Caching==1.3.3
Flask-Login==0.2.11
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.1
flask-swagger==0.2.13
Flask-WTF==0.14.2
funcsigs==1.0.0
future==0.16.0
gast==0.2.0
gitdb2==2.0.4
GitPython==2.1.11
google-auth==1.5.1
grpcio==1.15.0
gunicorn==19.9.0
h5py==2.8.0
html5lib==1.0.1
idna==2.7
ipykernel==4.9.0
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
iso8601==0.1.12
itsdangerous==0.24
jedi==0.12.1
Jinja2==2.8.1
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
kiwisolver==1.0.1
kubernetes==7.0.0
lockfile==0.12.2
lxml==3.8.0
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.3
mistune==0.8.3
monotonic==1.5
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==0.28.0
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.6.0
numpy==1.14.5
oauthlib==2.1.0
ordereddict==1.1
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pendulum==1.4.4
pexpect==4.6.0
pickleshare==0.7.4
plac==0.9.6
preshed==1.0.1
prometheus-client==0.3.1
prompt-toolkit==1.0.14
protobuf==3.6.1
protobuf3-to-dict==0.1.5
psutil==4.4.2
psycopg2-binary==2.7.5
ptyprocess==0.6.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
pyfiglet==0.7.5
Pygments==2.2.0
PyInquirer==1.0.2
PyJWT==1.6.4
pyparsing==2.2.1
python-daemon==2.1.2
python-dateutil==2.7.3
python-editor==1.0.3
python-nvd3==0.15.0
python-slugify==1.2.6
python3-openid==3.1.0
pytz==2018.5
pytzdata==2018.5
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.1
regex==2017.4.5
requests==2.19.1
requests-oauthlib==1.0.0
rsa==4.0
s3transfer==0.1.13
sagemaker==1.10.0
scikit-learn==0.19.2
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
setproctitle==1.1.10
simplegeneric==0.8.1
six==1.11.0
smmap2==2.0.4
spacy==2.0.12
SQLAlchemy==1.1.18
stoic==0.1
tabulate==0.7.7
tenacity==4.8.0
tensorboard==1.10.0
tensorflow==1.10.1
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
text-unidecode==1.2
thinc==6.10.3
thrift==0.9.3
toolz==0.9.0
tornado==5.1.1
tqdm==4.26.0
traitlets==4.3.2
tzlocal==1.5.1
ujson==1.35
unicodecsv==0.14.1
Unidecode==1.0.22
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.53.0
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wrapt==1.10.11
WTForms==2.2.1
zope.deprecation==4.3.0

